Refactor existing repeater into class
