+++ lib/maze/servlets/servlet.rb
@@ -28,7 +28,6 @@ module Maze
 
     # Receives and parses the requests and payloads sent from the test fixture
     class Servlet < BaseServlet
-      prepend RequestRepeater
 
       # Constructor
       #
@@ -40,6 +39,7 @@ module Maze
         @request_type = request_type
         @requests = Server.list_for request_type
         @schema = JSONSchemer.schema(schema) unless schema.nil?
+        @bugsnag_repeater = Maze::Repeaters::RequestRepeater.new(@request_type)
       end
 
       # Logs an incoming GET WEBrick request.
@@ -57,6 +57,9 @@ module Maze
       # @param request [HTTPRequest] The incoming GET request
       # @param response [HTTPResponse] The response to return
       def do_POST(request, response)
+
+        @bugsnag_repeater.repeat request
+
         # Turn the WEBrick HttpRequest into our internal HttpRequest delegate
         request = HttpRequest.new(request)
